lambda/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ndl-ocr-lite-lambda"
version = "1.0.0"
description = "NDL-OCR Lite Lambda handler"
requires-python = ">=3.10"
dependencies = [
    "onnxruntime==1.23.2",
    "pillow==12.1.1",
    "numpy==2.2.2",
    "PyYAML==6.0.1",
    "lxml==5.4.0",
    "networkx==3.3",
    "pyparsing==3.1.2",
    "ordered-set==4.1.0",
    "pypdfium2==4.30.0",
    "boto3",
]
